Bug fixed: Close button on each badge was not removing the item.
Issue: The issue arises from using the latest badge component from shadcn; which has a default behavior to set the pointer-events to none for SVG's.

Implementation: Same was fixed by applying  to the badge in the multi-select component.

Other Ways to solve the same (not-recommended): While there are other ways to solve the same problem, for example, changing the  in the badge component itself, but that would most probably break the components that you use from the shadcn itself. So it's wise to change that here.